requirements/test.txt
# Test dependencies go here.
-r base.txt

coverage==4.5.1
flake8==3.5.0 # pyup: != 2.6.0
django-test-plus==1.1.1
factory-boy==2.11.1
django-coverage-plugin==1.6.0

# pytest
pytest-django==3.4.2
pytest-sugar==0.9.1

# Selenium for UI testing
selenium==3.14.0

# BeautifulSoup for HTML testing
beautifulsoup4==4.6.3

# Black for formatting
black ==18.9b0

# Coverage through Codacy
codacy-coverage ==1.3.11
